fix(ui): address review feedback on parent app breadcrumb
- Fix +N dropdown navigation: use MenuItem action with
  navigation.goto() and a plain-string title instead of a nested
  <Link>, avoiding invalid nested anchors and stopPropagation
  swallowing the click.
- Make ancestor-walk cycle detection namespace-aware by keying the
  visited set on `${namespace}/${name}` so same-named apps in
  different namespaces are not treated as a cycle.
- Extract the duplicated ancestor-walk into a shared
  loadAppAncestors() helper in utils.tsx, reused by both the header
  breadcrumb and the SUMMARY PARENT APP row.
