+++ ui/src/app/applications/components/application-details/application-details.tsx
@@ -30,7 +30,7 @@ import {ApplicationResourceList, ApplicationResourceParentRef} from './applicati
 import {APPLICATION_DETAILS_SORT_KEY, ApplicationResourceSortKey, compareApplicationResource, GROUPED_NODES_DETAILS_SORT_KEY} from './application-resource-sort';
 import {useListSort} from '../../../shared/hooks/use-list-sort';
 import {Filters, FiltersProps, getEffectiveResourceFilter} from './application-resource-filter';
-import {getAppDefaultSource, getAppCurrentVersion, urlPattern, getAppParentName, getApplicationDetailsContainerClass} from '../utils';
+import {getAppDefaultSource, getAppCurrentVersion, urlPattern, getAppParentName, loadAppAncestors, getApplicationDetailsContainerClass} from '../utils';
 import {ChartDetails, OCIMetadata} from '../../../shared/models';
 import {ApplicationsDetailsAppDropdown} from './application-details-app-dropdown';
 import {useSidebarTarget} from '../../../sidebar/sidebar';
@@ -89,31 +89,14 @@ export const SelectNode = (fullName: string, containerIndex = 0, tab: string = n
 };
 
 const AppBreadcrumb = ({app, appName, objectListKind}: {app: appModels.Application; appName: string; objectListKind: string}) => {
+    const ctx = useContext(Context);
     const directParentName = getAppParentName(app);
     const appDropdown = <ApplicationsDetailsAppDropdown appName={appName} objectListKind={objectListKind} />;
     if (!directParentName) return appDropdown;
     return (
         <DataLoader
             input={`${app.metadata.name}/${app.metadata.namespace}`}
-            load={async () => {
-                const ancestors: Array<{name: string; namespace: string}> = [];
-                let currentName = directParentName;
-                let currentNamespace = app.metadata.namespace;
-                const visited = new Set<string>([app.metadata.name]);
-                for (let i = 0; i < 10; i++) {
-                    if (!currentName || visited.has(currentName)) break;
-                    ancestors.unshift({name: currentName, namespace: currentNamespace});
-                    visited.add(currentName);
-                    try {
-                        const parentApp = (await services.applications.get(currentName, currentNamespace, 'application')) as appModels.Application;
-                        currentName = getAppParentName(parentApp);
-                        currentNamespace = parentApp.metadata.namespace;
-                    } catch {
-                        break;
-                    }
-                }
-                return ancestors;
-            }}>
+            load={() => loadAppAncestors(app, directParentName)}>
             {(ancestors: Array<{name: string; namespace: string}>) => {
                 const chain = ancestors.filter(a => a.name !== app.metadata.name);
                 if (chain.length === 0) return appDropdown;
@@ -126,8 +109,8 @@ const AppBreadcrumb = ({app, appName, objectListKind}: {app: appModels.Applicati
                                 <DropDownMenu
                                     anchor={() => <a style={{cursor: 'pointer'}}>+{higher.length}</a>}
                                     items={higher.map(a => ({
-                                        title: <Link to={`/applications/${a.namespace}/${a.name}`}>{a.name}</Link>,
-                                        action: () => {}
+                                        title: a.name,
+                                        action: () => ctx.navigation.goto(`/applications/${a.namespace}/${a.name}`)
                                     }))}
                                 />
                                 <span style={{opacity: 0.5}}>›</span>
